backend/requirements.txt
fastapi[standard]>=0.135.0,<0.136.0
uvicorn[standard]>=0.34.0,<0.40.0
sqlmodel>=0.0.38,<0.1.0
alembic>=1.18.0,<1.19.0
psycopg2-binary>=2.9.10,<3.0.0
pyjwt>=2.10.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
# Pin bcrypt to 4.x — bcrypt 5.x removed __about__ and added strict length checks
# that break passlib 1.7.4 (known upstream incompatibility until passlib 1.8+).
bcrypt>=4.0.0,<5.0.0
pydantic-settings>=2.8.0,<3.0.0
